Asynchronous pagination
Still ~25 sec for one run on my machine. Now I understand a bit better
the case for a cache though I am still not convinced. The cache would:
- decouple probably a bit better IO from reporting logic
- pay off from the second run onward with significant performance
  improvement
- help work around request limits on Marvel API

Yet for me, the biggest value of such a CLI is its ability to stream
content and pipe it to other commands like `grep`. Where I stand this
outweighs the benefits of the cache as running it several times
consecutively is not an identified use case.